Move the Winters split into an `AvailablePotentialEnergyEquation` module
`BackgroundPotentialEnergy`, `AvailablePotentialEnergy` and the
`sorted_reference_height` machinery behind them now live in their own module
rather than inside `PotentialEnergyEquation`. The sorting is a whole-domain
operation with its own operand type, workspaces, sorting methods and grid
constraints, and it has little in common with the pointwise `PotentialEnergy`
kernel it was sharing a file with.

`PotentialEnergyEquation` keeps `PotentialEnergy` along with the
buoyancy-formulation type aliases and `validate_gravity_unit_vector`, which the
new module imports; it is included after `PotentialEnergyEquation` the same way
`SubFilterKineticEnergyEquation` builds on `FilteredKineticEnergyEquation`.

The tests split along the same line into a new `ape_diagnostics` group with its
own CI job (92 potential-energy tests stay, 328 available-potential-energy tests
move, which is the whole of the previous 420), and the docs into an
`available_potential_energy_equation.md` page. Every user-facing name is still
exported from `Oceanostics`, so nothing changes for callers.
