+++ CLAUDE.md
@@ -16,7 +16,7 @@ julia --project -e 'using Pkg; Pkg.test()'
 TEST_GROUP=vel_diagnostics julia --project -e 'using Pkg; Pkg.test()'
 ```
 
-Available TEST_GROUP values: `vel_diagnostics`, `tracer_diagnostics`, `u_momentum_diagnostics`, `v_momentum_diagnostics`, `w_momentum_diagnostics`, `ke_diagnostics`, `filtered_ke_diagnostics`, `subfilter_ke_diagnostics`, `tke_diagnostics`, `pe_diagnostics`, `active_tracer_diagnostics`, `tracer_variance_diagnostics`, `general_flow_diagnostics`, `canonical_flows`, `progress_messengers`, `spatial_filters`, `perf_invariants`.
+Available TEST_GROUP values: `vel_diagnostics`, `tracer_diagnostics`, `u_momentum_diagnostics`, `v_momentum_diagnostics`, `w_momentum_diagnostics`, `ke_diagnostics`, `filtered_ke_diagnostics`, `subfilter_ke_diagnostics`, `tke_diagnostics`, `pe_diagnostics`, `ape_diagnostics`, `active_tracer_diagnostics`, `tracer_variance_diagnostics`, `general_flow_diagnostics`, `canonical_flows`, `progress_messengers`, `spatial_filters`, `perf_invariants`.
 
 ```bash
 # Instantiate/build the package
@@ -50,17 +50,21 @@ All kernel functions use Oceananigans' staggered grid conventions with location 
 - **`TurbulentKineticEnergyEquation`**: TKE, isotropic dissipation, shear production rates (X/Y/Z and total)
 - **`TracerVarianceEquation`**: Tendency, dissipation rate, diffusion of tracer variance
 - **`PotentialEnergyEquation`**: Potential energy for BuoyancyTracer, linear/nonlinear SeawaterBuoyancy
-  (`PotentialEnergy`, Eₚ = -bz), plus the Winters et al. (1995) split into `BackgroundPotentialEnergy`
-  (E_b = -bz✶) and `AvailablePotentialEnergy` (Eₐ = -b(z - z✶)). The reference height z✶ comes from
-  `sorted_reference_height`, which returns a `Field` whose operand is a `SortedReferenceState` rather
-  than a `KernelFunctionOperation`: sorting is a whole-domain operation, so it hooks into `compute!`
-  the way `Oceananigans.Fields.Scan` does for `Integral`/`Average`, and is re-sorted on every
-  `compute!` so the diagnostic tracks the flow when written out during a simulation. The `method`
-  keyword picks one of three `AbstractSortingMethod`s, which agree on every volume integral:
-  `ThreeDimensionalSort` (default; each cell gets its own slot, so tied cells spread over a grid cell),
-  `HeavisideIntegral` (Winters eq. 11, tied cells share their layer's mid-height, so z✶ is a function
-  of buoyancy alone and local maps are clean), and `OneDimensionalSort` (the sorted column on its own
-  `1×1×N` grid of equal-volume cells; needs uniform cell volumes)
+  (`PotentialEnergy`, Eₚ = -bz). It owns the buoyancy-formulation type aliases
+  (`BuoyancyTracerModel`, `BuoyancyLinearEOSModel`, `BuoyancyBoussinesqEOSModel`, …) and
+  `validate_gravity_unit_vector`, which `AvailablePotentialEnergyEquation` imports.
+- **`AvailablePotentialEnergyEquation`**: the Winters et al. (1995) split of Eₚ into
+  `BackgroundPotentialEnergy` (E_b = -bz✶) and `AvailablePotentialEnergy` (Eₐ = -b(z - z✶)). The
+  reference height z✶ comes from `sorted_reference_height`, which returns a `Field` whose operand is a
+  `SortedReferenceState` rather than a `KernelFunctionOperation`: sorting is a whole-domain operation,
+  so it hooks into `compute!` the way `Oceananigans.Fields.Scan` does for `Integral`/`Average`, and is
+  re-sorted on every `compute!` so the diagnostic tracks the flow when written out during a simulation.
+  The `method` keyword picks one of three `AbstractSortingMethod`s, which agree on every volume
+  integral: `ThreeDimensionalSort` (default; each cell gets its own slot, so tied cells spread over a
+  grid cell), `HeavisideIntegral` (Winters eq. 11, tied cells share their layer's mid-height, so z✶ is a
+  function of buoyancy alone and local maps are clean), and `OneDimensionalSort` (the sorted column on
+  its own `1×1×N` grid of equal-volume cells; needs uniform cell volumes). Built on
+  `PotentialEnergyEquation`, so it is included after it
 - **`FlowDiagnostics`**: Richardson/Rossby numbers, Ertel/ThermalWind potential vorticity, strain rate & vorticity tensor moduli, Q-criterion, `subfilter_covariance` (generalized subfilter covariance `τ(a,b) = filter(a·b) − filter(a)·filter(b)`, unifying subfilter tracer flux and momentum stress), MixedLayerDepth, BottomCellValue
 - **`ProgressMessengers`** (submodule): Composable simulation progress reporters using `+` (comma-separated) and `*` (concatenation) operators
