+++ docs/make.jl
@@ -111,6 +111,7 @@ pages = ["Home" => "index.md",
                                 "Turbulent kinetic energy equation"      => "turbulent_kinetic_energy_equation.md",
                                 "Tracer variance equation"               => "tracer_variance_equation.md",
                                 "Potential energy equation"              => "potential_energy_equation.md",
+                                "Available potential energy equation"    => "available_potential_energy_equation.md",
                                 ],
          "Flow diagnostics" => "flow_diagnostics.md",
          "Progress messengers" => "progress_messengers.md",
